+++ backend/BrewInventory.Application/Contracts/Recipes/CreateRecipeFermentableRequest.cs
@@ -1,3 +1,3 @@
 namespace BrewInventory.Application.Contracts.Recipes;
 
-public record CreateRecipeFermentableRequest(int FermentableId, decimal Amount);
+public record CreateRecipeFermentableRequest(int FermentableId, double Amount);
