+++ backend/BrewInventory.Application/Contracts/Recipes/CreateRecipeHopRequest.cs
@@ -1,3 +1,3 @@
 namespace BrewInventory.Application.Contracts.Recipes;
 
-public record CreateRecipeHopRequest(int HopId, decimal Amount);
+public record CreateRecipeHopRequest(int HopId, double Amount);
