+++ backend/BrewInventory.Application/Contracts/Recipes/CreateRecipeMiscRequest.cs
@@ -1,3 +1,3 @@
 namespace BrewInventory.Application.Contracts.Recipes;
 
-public record CreateRecipeMiscRequest(int MiscId, decimal Amount);
+public record CreateRecipeMiscRequest(int MiscId, double Amount);
