+++ backend/BrewInventory.Application/Contracts/Recipes/CreateRecipeYeastRequest.cs
@@ -1,3 +1,3 @@
 namespace BrewInventory.Application.Contracts.Recipes;
 
-public record CreateRecipeYeastRequest(int YeastId, decimal Amount);
+public record CreateRecipeYeastRequest(int YeastId, double Amount);
